refactor(css): allow style component declaration in css files
