+++ package.json
@@ -35,6 +35,7 @@
     "@total-typescript/ts-reset": "^0.4.2",
     "@types/cookie": "^0.5.1",
     "@types/gtag.js": "^0.0.12",
+    "@types/postcss-js": "^4.0.0",
     "@types/postcss-mixins": "^9.0.0",
     "@typescript-eslint/eslint-plugin": "^5.56.0",
     "@typescript-eslint/parser": "^5.56.0",
@@ -52,6 +53,7 @@
     "postcss": "^8.4.21",
     "postcss-html": "^1.5.0",
     "postcss-import": "^15.1.0",
+    "postcss-js": "^4.0.1",
     "postcss-load-config": "^4.0.1",
     "postcss-mixins": "^9.0.4",
     "postcss-nesting": "^11.2.2",
