+++ src/lib/client/styles/components/avatar.css
@@ -1,0 +1,8 @@
+.c-avatar {
+  aspect-ratio: 1 / 1;
+  min-width: 40px;
+  min-height: 40px;
+
+  background-color: theme('colors.bg.300');
+  border-radius: theme('borderRadius.full');
+}
