+++ src/lib/client/styles/components/btn.css
@@ -1,0 +1,59 @@
+.c-btn {
+  display: grid;
+  column-gap: 8px;
+  align-items: center;
+  justify-content: center;
+
+  min-height: 40px;
+  padding: 0 12px;
+
+  border-radius: theme('borderRadius.DEFAULT');
+
+  &:where(:has(:is(img,svg))) {
+    grid-template-columns: auto auto;
+  }
+
+  &:where(:not(:has(:is(img,svg)))) {
+    text-align: center;
+  }
+
+  &:where(:not(:is(.c-btn--outlined, .c-btn--text))),
+  &:where(.c-btn--default) {
+    white-space: nowrap;
+    background-color: theme('colors.bg.300');
+  }
+
+  &:where(.c-btn--outlined) {
+    border: 1px solid theme('colors.bg.300');
+  }
+
+  &:where(.c-btn--text) {
+    position: relative;
+    padding: 0 0 4px;
+    font-weight: 500;
+
+    &::after {
+      content: '';
+
+      position: absolute;
+      top: 100%;
+      left: 0;
+      transform-origin: right center;
+      transform: scaleX(0);
+
+      display: block;
+
+      width: 100%;
+      height: 1;
+
+      background: theme('colors.fg.200');
+
+      transition: transform 250ms theme('transitionTimingFunction.DEFAULT');
+    }
+
+    &:hover::after {
+      transform-origin: left center;
+      transform: scaleX(1);
+    }
+  }
+}
