+++ src/lib/client/styles/components/container.css
@@ -1,0 +1,12 @@
+.c-container {
+  max-width: 1200px;
+  margin-right: auto;
+  margin-left: auto;
+  padding-right: 24px;
+  padding-left: 24px;
+
+  @screen md {
+    padding-right: 32px;
+    padding-left: 32px;
+  }
+}
