+++ src/lib/client/styles/components/graphic.css
@@ -1,0 +1,8 @@
+.c-graphic {
+  aspect-ratio: 1 / 1;
+  min-width: 120px;
+  min-height: 120px;
+
+  background-color: theme('colors.bg.300');
+  border-radius: theme('borderRadius.DEFAULT');
+}
